requirements.txt
python-dotenv
discord.py==2.4.0
